Update unit tests and benchmark
